User subscription, subscription, follow features changed

+++ src/main/java/ru/ermakov/creator/model/Subscription.java
@@ -6,6 +6,6 @@ public record Subscription(
         Creator creator,
         String title,
         String description,
-        Integer price
+        Long price
 ) {
 }
