+++ src/main/java/ru/ermakov/creator/model/SubscriptionRequest.java
@@ -5,6 +5,6 @@ public record SubscriptionRequest(
         String creatorId,
         String title,
         String description,
-        Integer price
+        Long price
 ) {
 }
